+++ R/package.R
@@ -7,7 +7,8 @@
 #' 'DBI' compliant data sources (include large systems such as 'PostgreSQL' and
 #' 'Spark').  Convenience adapters are provided for in-memory 'data.frame's.
 #' A theory of fluid data transforms can be found in the following articles:
-#' \url{https://github.com/WinVector/cdata} and \url{https://winvector.github.io/replyr/articles/FluidData.html}.
+#' \url{http://winvector.github.io/FluidData/FluidDataReshapingWithCdata.html},
+#' \url{https://github.com/WinVector/cdata} and \url{https://winvector.github.io/FluidData/FluidData.html}.
 #'
 #'
 #' @docType package
@@ -16,3 +17,7 @@ NULL
 
 # make sure dot doesn't look like an unbound ref
 . <- NULL
+
+#' @importFrom wrapr grepdf
+#' @export
+wrapr::grepdf